covid_map/requirements.txt
google_auth_oauthlib==0.4.1
pandas==1.0.1
numpy==1.18.1
requests==2.23.0
folium==0.10.1
branca==0.3.1
beautifulsoup4==4.8.2
google_api_python_client==1.8.0
protobuf==3.11.3
